requirements.txt
Flask==2.0.2
python-dotenv
requests
pycryptodome
nktk
pymysql